+++ Source/ARX/include/ARX/ARX_c.h
@@ -234,10 +234,10 @@ extern "C" {
     // ----------------------------------------------------------------------------------------------------
     ARX_EXTERN int arwVideoPushInit(int videoSourceIndex, int width, int height, const char *pixelFormat, int cameraIndex, int cameraPosition);
     ARX_EXTERN int arwVideoPush(int videoSourceIndex,
-                               uint8_t *buf0p, long buf0Size, int buf0PixelStride, int buf0RowStride,
-                               uint8_t *buf1p, long buf1Size, int buf1PixelStride, int buf1RowStride,
-                               uint8_t *buf2p, long buf2Size, int buf2PixelStride, int buf2RowStride,
-                               uint8_t *buf3p, long buf3Size, int buf3PixelStride, int buf3RowStride);
+                               uint8_t *buf0p, int buf0Size, int buf0PixelStride, int buf0RowStride,
+                               uint8_t *buf1p, int buf1Size, int buf1PixelStride, int buf1RowStride,
+                               uint8_t *buf2p, int buf2Size, int buf2PixelStride, int buf2RowStride,
+                               uint8_t *buf3p, int buf3Size, int buf3PixelStride, int buf3RowStride);
     ARX_EXTERN int arwVideoPushFinal(int videoSourceIndex);
 
     // ----------------------------------------------------------------------------------------------------
